Added: material densities and ToString method to Distance Distributions

+++ Graph/DistributionData.cs
@@ -1,7 +1,7 @@
 ﻿namespace Graph {
     public class DistributionData {
         public List<float> Distances = new();
-        public List<float> MaterialDensities = new();
+        public List<double> MaterialDensities = new();
         public List<double> ProbabilityDensities = new();
         public List<double> CummulativeProbabilities = new();
     }
